refactor: set playlist as a commom struct
